Rename: address review comments
